==> notebooks/1_estructura.ipynb <==
import pandas as pd

# CARGAMOS TABLAS
df_trending = pd.read_csv('../data/trending.csv')
df_reviews = pd.read_csv('../data/reviews.csv')

df_trending.head()

df_reviews.head()

# AJUSTAMOS NOMBRES DE LAS COLUMNAS
df_trending.columns = ['rank', 'book_title', 'book_price', 'rating', 'author', 'year_of_publication', 'genre', 'url']
df_reviews.columns = ['sno', 'book_name', 'review_title', 'reviewer', 'reviewer_rating', 'review_description', 'is_verified', 'date', 'timestamp', 'asin']

# UNIMOS LAS TABLAS
df_datos = pd.merge(df_trending, df_reviews, left_on='book_title', right_on='book_name', how='inner')
df_datos.head()

# ELIMINAMOS COLUMNA REPETIDA
df_datos = df_datos.drop('book_name', axis=1)
df_datos = df_datos.drop('sno', axis=1)

# EXPORTAMOS EL DF DE DATOS
df_datos.to_csv('../data/datos.csv', index=False)

filas, columnas = df_datos.shape
print(f"Número de filas: {filas}")
print(f"Número de columnas: {columnas}")
