==> notebooks/2_procesamiento.ipynb <==
import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer
from nltk.corpus import stopwords
from sklearn.preprocessing import StandardScaler

# DIVIDIMOS EL DATASET EN TRAIN Y TEST
train = pd.read_csv('../data/datos.csv')

# CARGAR STOP WORDS
stop_words = stopwords.words('english')

train.head()

# Procesamiento de Train
train = train.drop(columns=['rank', 'url', 'timestamp', 'asin', 'author', 'year_of_publication', 'date', 'asin', 'book_price','is_verified'])

# Separar los géneros
train['genre'] = train['genre'].str.split(',')
train = train.explode('genre')
train = pd.get_dummies(train, columns=['genre'], drop_first=False)

# Procesamiento de las reseñas de los usuarios con TF-IDF ajustado
tfidf_vectorizer_title = TfidfVectorizer(max_features=100, stop_words=stop_words, max_df=0.8, min_df=5)
tfidf_vectorizer_description = TfidfVectorizer(max_features=100, stop_words=stop_words, max_df=0.8, min_df=5)

# Aplicar TF-IDF
tfidf_title = tfidf_vectorizer_title.fit_transform(train['review_title'].fillna(''))
tfidf_description = tfidf_vectorizer_description.fit_transform(train['review_description'].fillna(''))

# Convertir TF-IDF en DataFrames con prefijos
tfidf_title_df = pd.DataFrame(tfidf_title.toarray(), columns=[f'word_title_{word}' for word in tfidf_vectorizer_title.get_feature_names_out()], index=train.index)
tfidf_description_df = pd.DataFrame(tfidf_description.toarray(), columns=[f'word_desc_{word}' for word in tfidf_vectorizer_description.get_feature_names_out()], index=train.index)

# Eliminar columnas de TF-IDF con palabras de longitud menor a 3
tfidf_title_df = tfidf_title_df.loc[:, [col for col in tfidf_title_df.columns if len(col) > len("word_title_") + 2]]
tfidf_description_df = tfidf_description_df.loc[:, [col for col in tfidf_description_df.columns if len(col) > len("word_desc_") + 2]]

# Concatenarlos
train = pd.concat([train, tfidf_title_df, tfidf_description_df], axis=1)

# Eliminar columnas originales
train.drop(columns=['review_title', 'review_description'], inplace=True)

train.head()

# EXPORTAMOS LOS DF
train.to_csv('../data/train.csv', index=False)
